+++ cpclib-bndbuild/src/lib.rs
@@ -31,7 +31,9 @@ use thiserror::Error;
 pub use crate::BndBuilder;
 use crate::event::BndBuilderObserverRc;
 use crate::executor::*;
-use crate::task::{BASMDOC_CMDS, CATALOG_CMDS, CPC2IMG_CMDS, GRAFX2_CMDS, LOCOMOTIVE_CMDS, MINY_CMDS, RTZX_CMDS};
+use crate::task::{
+    BASMDOC_CMDS, CATALOG_CMDS, CPC2IMG_CMDS, GRAFX2_CMDS, LOCOMOTIVE_CMDS, MINY_CMDS, RTZX_CMDS
+};
 
 pub mod app;
 pub mod builder;