+++ CHANGELOG.md
@@ -14,15 +14,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `cpclib-catart`add this crate to handle catalog art
 - `cpclib-csl`add support for CSL file parsing and generation (mainly to check validity of existing ones)
 - `cpclib-basic` add support for binary encoded programs (tokenized BASIC)
+- `cpclib-basmdoc` add a new crate to handle documetnation of z80 projects
 - `cpclib-bndbuild` add support fof Z80Profiler by Targhan/Arkos
 - `cpclib-bndbuild` add support of the catalog command
 - `cpclib-emucontrol` add support to activate roms (it was only possible to dectivate them before)
+- `cpclib-locomotive` new crate to handle the executable for basisc manipulation
 - `cpclib-orgams-ascii` add support to ORGAMS files. This crate aims at converting orgams sourceode to ascii and ascii source code to orgams. (in fact utf8, but...)
 
 ### Changed
 - Standardized README filename from `.mkd` to `.md`
 - Improved workspace dependency management
 - `cpclib-basic` better support of string programs
+- `cpclib-bndbuild` add support to catalog, locomotive, csl, basmdoc
 - `cpclib-catalog` add catalog visualization and catart creation
 - `cpclib-basm` add reorganize the source cdode of the parser
 - `cpclib-emucontrol` add support of CSL for controling emulators (partially possible for those without CSL support)